# public/openapi.yaml
openapi: 3.1.0
info:
  title: TrustForge AI API
  version: 0.1.0
  description: Evidence-backed trust registry API. This contract will expand without moving domain logic into the transport layer.
servers:
  - url: /api/v1
paths:
...
  /evidence:
    get:
      operationId: listEvidence
      summary: List evidence for a registry subject
      parameters:
        - { in: query, name: subjectType, required: true, schema: { type: string } }
        - { in: query, name: subjectId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Evidence records in observation order }
    post:
      operationId: submitEvidence
      summary: Submit versioned evidence
      responses:
        "201": { description: Evidence submitted for adjudication }
...